# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "مسیر کانکور — API",
    "description": "API بک‌اند سایت آموزشگاه کانکور. اکثر مسیرها نیازمند توکن JWT هستند (هدر Authorization: Bearer <token>).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "نسبی — همان دامنه‌ی فرانت‌اند (پشت پروکسی/nginx)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
...
      "Course": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "subject": { "type": "string" },
          "branch_id": { "type": "integer", "nullable": true },
          "teacher_id": { "type": "integer", "nullable": true },
          "price": { "type": "integer", "description": "قیمت به افغانی؛ ۰ یعنی رایگان" },
          "price_usd": { "type": "number", "nullable": true, "description": "قیمت جدا به دلار برای پرداخت با کارت (Stripe). اگر خالی باشد، پرداخت با کارت برای این کورس غیرفعال است." },
          "cover_color": { "type": "string" },
          "branch_title": { "type": "string" },
          "teacher_name": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "توکن ارسال نشده یا نامعتبر است.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "نقش کاربر اجازه‌ی این عملیات را نمی‌دهد.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "منبع مورد نظر یافت نشد.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/courses": {
      "get": {
        "tags": ["courses"],
        "summary": "لیست عمومی کورس‌ها",
        "security": [],
        "parameters": [{ "name": "branch_id", "in": "query", "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Course" } } } } } }
      },
      "post": {
        "tags": ["courses"],
        "summary": "ساخت کورس جدید (استاد/ادمین)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "subject": { "type": "string" },
                  "branch_id": { "type": "integer" },
                  "teacher_id": { "type": "integer", "description": "فقط ادمین می‌تواند تعیین کند" },
                  "price": { "type": "integer" },
                  "cover_color": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "ساخته شد" },
          "400": { "description": "ورودی نامعتبر" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/courses/mine": {
      "get": {
        "tags": ["courses"],
        "summary": "کورس‌های خود استاد (یا همه‌ی کورس‌ها برای ادمین)",
        "responses": { "200": { "description": "OK" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/courses/{id}": {
      "get": {
        "tags": ["courses"],
        "summary": "جزئیات یک کورس (بدون URL ویدیو/پاسخ آزمون)",
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/courses/{id}/enrollment": {
      "get": {
        "tags": ["courses"],
        "summary": "وضعیت ثبت‌نام/دسترسی کاربر جاری به این کورس",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/courses/{id}/students": {
      "get": {
        "tags": ["courses"],
        "summary": "لیست شاگردان ثبت‌نام‌شده (استاد/ادمین)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
    "/forum/course/{courseId}": {
      "get": {
        "tags": ["forum"],
        "summary": "موضوعات انجمن یک کورس",
        "parameters": [{ "name": "courseId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["forum"],
        "summary": "ساخت موضوع جدید در انجمن کورس",
        "parameters": [{ "name": "courseId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["title"], "properties": { "title": { "type": "string", "maxLength": 200 }, "body": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "ساخته شد" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
  }
}